LPS-125991 Rename CommerceCountryHelper to CommerceCountryManager
