+++ modules/apps/commerce/commerce-api/bnd.bnd
@@ -6,6 +6,7 @@ Export-Package:\
 	com.liferay.commerce.configuration,\
 	com.liferay.commerce.constants,\
 	com.liferay.commerce.context,\
+	com.liferay.commerce.country,\
 	com.liferay.commerce.discount,\
 	com.liferay.commerce.exception,\
 	com.liferay.commerce.health.status,\
