+++ modules/apps/commerce/commerce-api/src/main/resources/com/liferay/commerce/country/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
